chore: 프로필에 따라 인가 on/off 기능 추가
