+++ src/main/java/upbrella/be/config/WebConfig.java
@@ -1,11 +1,8 @@
 package upbrella.be.config;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import upbrella.be.config.interceptor.OAuthLoginInterceptor;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
@@ -19,24 +16,4 @@ public class WebConfig implements WebMvcConfigurer {
                 .allowedMethods("*")
                 .allowedHeaders("*");
     }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(oAuthLoginInterceptor())
-                .addPathPatterns("/**")
-                .excludePathPatterns(
-                        "/users/login/**",
-                        "/users/join/**",
-                        "/index.html",
-                        "/error/**",
-                        "/api/error/**",
-                        "/docs/**");
-    }
-
-    @Bean
-    public OAuthLoginInterceptor oAuthLoginInterceptor() {
-        return new OAuthLoginInterceptor();
-    }
-
-
 }
