+++ src/main/resources/application.yml
@@ -1,22 +1,3 @@
 spring:
-  config:
-    import: "classpath:application.properties"
-  datasource:
-    url: ${DATABASE_URL}
-    username: ${DATABASE_USERNAME}
-    password: ${DATABASE_PASSWORD}
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    properties:
-      hibernate:
-        format_sql: true
-
-    show-sql: true
-logging:
-  level:
-    org:
-      hibernate:
-        SQL: DEBUG
-        type:
-          descriptor:
-            sql: trace
+  profiles:
+    active: prod
